+++ tests/r_parity/test_gintervals_intersect.py
@@ -33,7 +33,7 @@ _CASES = {
     "gintervals.intersect.2": (lambda: pm.gintervals_intersect(_rb(_scr("test.fixedbin > 0.1 & test.fixedbin < 0.3", _i1([1, 2])), _scr("test.fixedbin > 0.2 & test.fixedbin < 0.4", _i1([1, 2]))), _scr("test.fixedbin > 0.34 & test.fixedbin < 0.5", _i1([1, 2]))), None),
     "gintervals.intersect.3": (lambda: pm.gintervals_intersect(_scr("test.rects > 40", _i2([1, 2, 5, 8], 0, -1)), _scr("test.rects < 60", _i2([2, 4, 5, 9], 0, -1))), GAP_2D),
     "gintervals.intersect.named.1": (lambda: pm.gintervals_intersect("test.bigintervs_1d_1", "test.bigintervs_1d_2"), None),
-    "gintervals.intersect.named.2": (lambda: pm.gintervals_intersect("test.generated_1d_1", "test.generated_1d_2"), GAP_NAMED_TRACK),
+    "gintervals.intersect.named.2": (lambda: pm.gintervals_intersect("test.generated_1d_1", "test.generated_1d_2"), None),
     "gintervals.intersect.named.3": (lambda: pm.gintervals_intersect("test.bigintervs_2d_5", "test.bigintervs_2d_6"), GAP_2D),
     "gintervals.intersect.named.4": (lambda: pm.gintervals_intersect("test.generated_2d_5", "test.generated_2d_6"), GAP_2D),
     "gintervals.union.1": (lambda: pm.gintervals_union(_scr("test.fixedbin > 0.1 & test.fixedbin < 0.3", _i1([1, 2], 0, -1)), _scr("test.fixedbin < 0.2", _i1([1, 2], 0, -1))), None),